+++ kwsparentsdk/src/main/java/kws/superawesome/tv/kwsparentsdk/services/create/KWSCreateParentInterface.java
@@ -1,0 +1,5 @@
+package kws.superawesome.tv.kwsparentsdk.services.create;
+
+public interface KWSCreateParentInterface {
+    void didCreateParent (KWSCreateParentStatus status);
+}