+++ kwsparentsdk/src/main/java/kws/superawesome/tv/kwsparentsdk/services/create/KWSCreateParentStatus.java
@@ -1,0 +1,13 @@
+package kws.superawesome.tv.kwsparentsdk.services.create;
+
+/**
+ * Created by gabriel.coman on 22/12/2016.
+ */
+
+public enum KWSCreateParentStatus {
+    CREATED,
+    DUPLICATE,
+    INVALID_EMAIL,
+    INVALID_PASSWORD,
+    NETWORK_ERROR
+}