+++ kwsparentsdk/src/main/java/kws/superawesome/tv/kwsparentsdk/services/kids/KWSGetKidsServiceInterface.java
@@ -1,9 +1,0 @@
-package kws.superawesome.tv.kwsparentsdk.services.kids;
-
-import java.util.List;
-
-import kws.superawesome.tv.kwsparentsdk.models.kids.KWSChild;
-
-public interface KWSGetKidsServiceInterface {
-    void gotChildren (List<KWSChild> children);
-}