+++ kwsparentsdk/src/main/java/kws/superawesome/tv/kwsparentsdk/services/oauth/KWSAuthInterface.java
@@ -2,6 +2,6 @@ package kws.superawesome.tv.kwsparentsdk.services.oauth;
 
 import kws.superawesome.tv.kwsparentsdk.models.oauth.KWSLoggedUser;
 
-public interface KWSAuthServiceInterface {
+public interface KWSAuthInterface {
     void didAuthUser (KWSLoggedUser user);
 }