+++ kwsparentsdk/src/main/java/kws/superawesome/tv/kwsparentsdk/services/parent/KWSGetParentInterface.java
@@ -2,6 +2,6 @@ package kws.superawesome.tv.kwsparentsdk.services.parent;
 
 import kws.superawesome.tv.kwsparentsdk.models.parent.KWSParentUser;
 
-public interface KWSGetParentServiceInterface {
+public interface KWSGetParentInterface {
     void gotParent (KWSParentUser parent);
 }