+++ kwsparentsdk/src/main/java/kws/superawesome/tv/kwsparentsdk/services/parent/KWSUpdateParentInterface.java
@@ -4,6 +4,6 @@ package kws.superawesome.tv.kwsparentsdk.services.parent;
  * Created by gabriel.coman on 21/12/2016.
  */
 
-public interface KWSUpdateParentServiceInterface {
+public interface KWSUpdateParentInterface {
     void updatedParent (boolean success);
 }